+++ docarray/typing/__init__.py
@@ -1,10 +1,12 @@
 from docarray.typing.id import ID
-from docarray.typing.tensor import NdArray, Tensor, TorchEmbedding, TorchTensor
+from docarray.typing.tensor import NdArray, Tensor
+
+# from docarray.typing.tensor import TorchEmbedding, TorchTensor
 from docarray.typing.tensor.embedding import Embedding
 from docarray.typing.url import AnyUrl, ImageUrl, TextUrl
 
 __all__ = [
-    'TorchTensor',
+    # 'TorchTensor',
     'NdArray',
     'Embedding',
     'ImageUrl',
@@ -12,5 +14,5 @@ __all__ = [
     'AnyUrl',
     'ID',
     'Tensor',
-    'TorchEmbedding',
+    # 'TorchEmbedding',
 ]
