+++ docarray/typing/tensor/__init__.py
@@ -1,13 +1,17 @@
-from docarray.typing.tensor.embedding import Embedding, NdArrayEmbedding, TorchEmbedding
+from docarray.typing.tensor.embedding import (  # , TorchEmbedding
+    Embedding,
+    NdArrayEmbedding,
+)
 from docarray.typing.tensor.ndarray import NdArray
 from docarray.typing.tensor.tensor import Tensor
-from docarray.typing.tensor.torch_tensor import TorchTensor
+
+# from docarray.typing.tensor.torch_tensor import TorchTensor
 
 __all__ = [
     'NdArray',
-    'TorchTensor',
+    # 'TorchTensor',
     'Tensor',
     'Embedding',
-    'TorchEmbedding',
+    # 'TorchEmbedding',
     'NdArrayEmbedding',
 ]
