+++ docarray/typing/tensor/embedding/__init__.py
@@ -1,0 +1,5 @@
+# from docarray.typing.tensor.embedding.embedding_torch import TorchEmbedding
+from docarray.typing.tensor.embedding.embedding import Embedding
+from docarray.typing.tensor.embedding.embedding_ndarray import NdArrayEmbedding
+
+__all__ = ['NdArrayEmbedding', 'Embedding']
