+++ docarray/typing/tensor/embedding/embedding.py
@@ -1,0 +1,7 @@
+from typing import Union
+
+from docarray.typing.tensor.embedding.embedding_ndarray import NdArrayEmbedding
+
+# from docarray.typing.tensor.embedding.embedding_torch import TorchEmbedding
+
+Embedding = Union[NdArrayEmbedding]  # , TorchEmbedding]
